refactor(contact): import ColumnType from unify.common.column_types
Mechanical follow-up after extracting ColumnType out of KnowledgeManager so
contact field typing and search tests no longer depend on the deleted KM types
module path.
